requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
litellm>=1.51.0
traceloop-sdk>=0.28.0
python-dotenv>=1.0.0
pydantic-settings>=2.5.0
httpx>=0.27.0
